--- src/pages/liveChat/fixedChat/fixedChat.css
@@ -1,24 +1,0 @@
-.message {
-
-    background: var(--buttonBg);
-    color: white;
-    padding: 0.7rem;
-    border-radius: 1rem 1rem 1rem 0;
-    max-width: 28rem;
-    width: fit-content;
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-}
-
-.message>span:nth-child(2) {
-    font-size: 0.7rem;
-    color: var(--textColor);
-    align-self: end;
-}
-
-.own {
-    align-self: flex-end;
-    border-radius: 1rem 1rem 0 1rem;
-    background: linear-gradient(98.63deg, #24e4f0 0%, #358ff9 100%);
-}
