+++ src/pages/liveChat/massage/Massage.jsx
@@ -1,13 +1,8 @@
-import SimpleChatBot from "../../chatBot/simpleChatBot/SimpleChatBot";
-import FixedChat from "../fixedChat/FixedChat";
+import SimpleChatBot from "../chatBot/chatBot/SimpleChatBot";
 
 const Massage = () => {
     return (
         <div className="relative h-10">
-            {/* Your main content */}
-
-            {/* Fixed Icon and Menu */}
-            {/* <FixedChat /> */}
             <SimpleChatBot />
         </div>
     );
